Renamed common_settings.py for brevity.
